requirements.txt
asgiref==3.8.1
black==24.8.0
bleach==6.1.0
click==8.1.7
crispy-bootstrap4==2024.1
Django==4.2.15
django-crispy-forms==2.3
django-js-asset==2.2.0
django-tinymce==4.1.0
flake8==7.1.1
gunicorn==23.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
psycopg==3.1.18
psycopg-binary==3.1.18
pycodestyle==2.12.1
pyflakes==3.2.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
webencodings==0.5.1
